Add files via upload
Sometimes this method turns out to be extra ordinarily quick even with very large numbers which depends on the numbers and the 2 integer settings as to if it is quick. 
I have set the settings to test random Integers out on the algorithm In an attempt to try and solve RSA 230.